Refactor: 상품 업로드 API 분리
